backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-celery-beat==2.5.0
gunicorn==21.2.0
python-dotenv==1.0.0